==> 1.eland_csv_to_tsv.ipynb <==
import pandas as pd

df_news = pd.read_csv("https://github.com/roccqqck/news_bert/raw/master/data/2015_Company.csv", encoding="utf-8")
df_news

df_news.shape

df_news['風險類型'].value_counts()

# column name rename "編號":"id", "風險類型": "label", "段落摘要": "text"
df_news = df_news.rename(columns={"編號":"id", "風險類型": "label", "段落摘要": "text"})
df_news.head(2)

# df_news['風險事件區間'] = pd.to_datetime(df_news['風險事件區間'])
df_news['text'] = df_news['text'].astype(str).str.replace('\t' , ' ').str.replace("\n", " ")    # tab都換成空格

# 只取特定欄位
df_news2 = df_news.loc[:, ['id', 'label', 'text']]
df_news2.head(2)

from snownlp import SnowNLP

def sentiments_all(text):
    s = SnowNLP(u""+text)       # 宣告物件 s = SnowNLP(u"这个东西真心很不错呀")
    text2 = s.han               # snownlp 要簡體中文效果比較好
    s = SnowNLP(u""+text2)
    sentiments_score = s.sentiments         # 得到 0 ~ 1的sentiments分數
    return sentiments_score

# text rank 演算法 死出top5的關鍵字
def keywords_all(text):
    s = SnowNLP(u""+text)       # 宣告物件 s = SnowNLP(u"这个东西真心很不错呀")
    text2 = s.han               # snownlp 要簡體中文效果比較好
    s = SnowNLP(u""+text2)
    top_list = s.keywords(5)        # 得到 top5的關鍵字
    return top_list

# text rank 演算法 死出top5的關鍵字
def text_add_keywords(text):
    s = SnowNLP(u""+text)       # 宣告物件 s = SnowNLP(u"这个东西真心很不错呀")
    text2 = s.han               # snownlp 要簡體中文效果比較好
    s = SnowNLP(u""+text2)
    top_list = s.keywords(5)     # top5的關鍵字
    text2 = "".join(top_list) + text   # top5的關鍵字加到文章的最前面
    return text2

text = "「陽光普照」一舉拿下最佳劇情長片、最佳導演獎、最佳男主角獎、最佳男配角獎、最佳剪輯獎、觀眾票選獎。"

sentiments_all(text)

top_list = keywords_all(text)    # 把文章丟進去 回傳top5的關鍵字
print(type(top_list))
top_list

text2 = "".join(top_list) + text  # top5的關鍵字加到文章的最前面
text2

text_add_keywords(text)

%%time
df_news2['keywords'] = df_news2['text'].apply(keywords_all)
df_news2['text2'] = df_news2['text'].apply(text_add_keywords)
df_news2['sentiments'] = df_news2['text'].apply(sentiments_all)   # 得到 0 ~ 1的分數
df_news2['sentiments'] = 2*(df_news2['sentiments'] - 0.5)             # 轉成 -1 ~ 1的分數
df_news2

text = "辭舊迎新之際，你卻在遺憾2018年沒有等到能夠打動自己的黑科技手機？"
print(len(text))
text[:10]  # 只取前10個字

# 文章字數 > 510了話 去尾
# 字數小於512-2 因為還有CLS SEP
def remove_510(text):
    if len(text) > 510:
        text = text[:510]    # 只取前510個字
    return text

# # 只留字數 <= 510的文章
# # 字數小於512-2 因為還有CLS SEP
# MAX_LENGTH = 510
# # df_news2 = df_news2[~(df_news2["text"].apply(lambda x : len(x)) > MAX_LENGTH)]
# df_news2 = df_news2[df_news2["text"].apply(lambda x : len(x) <= MAX_LENGTH) ]  # 去掉字數>510的文章
# df_news2 = df_news2.reset_index(drop=True)      # 重設index
# df_news2['id'] = df_news2.index                 # index 變成 id column
# df_news2

df_news2["text"] = df_news2["text"].apply(remove_510)
df_news2["text2"] = df_news2["text2"].apply(remove_510)
df_news2.head(2)

df_news2["label"].unique() 

df_news2.loc[df_news2['label'] == "財務營運風險", 'label2'] = 0
df_news2.loc[df_news2['label'] == "法律風險", 'label2'] = 1
df_news2.loc[df_news2['label'] == "資安風險", 'label2'] = 2
df_news2.loc[df_news2['label'] == "勞動風險", 'label2'] = 3
df_news2.loc[df_news2['label'] == "管理風險", 'label2'] = 4
df_news2.head(2)

df_news2["label2"].unique() 

df_news2["label"] = df_news2["label2"]
df_news2["label"] = df_news2["label"].astype(int)
df_news2 = df_news2.drop(columns=['label2'])
df_news2.head(2)

df_news3 = df_news2.drop(columns=['label2'])
df_news3.head(2)

# idempotence, 將處理結果另存成 tsv 供 PyTorch 使用   # header=False會沒有column名   # \t 是tab
df_news3.to_csv("data/2015_Company_textrank_sentiments.tsv", sep="\t", index=False)

df_news3 = df_news3.drop(columns=['sentiments', 'text2'])
df_news3.head(2)

# idempotence, 將處理結果另存成 tsv 供 PyTorch 使用   # header=False會沒有column名   # \t 是tab
df_news3.to_csv("data/2015_Company.tsv", sep="\t", index=False)

# # csv 用,區隔
# df_news2.to_csv("data/2015_Company.2.csv", index=False)

# 讀取tsv測試
df_news3 = pd.read_csv("data/2015_Company.tsv", sep="\t", encoding="utf-8")
df_news3

# 讀取tsv測試
df_news3 = pd.read_csv("data/2015_Company_textrank_sentiments.tsv", sep="\t", encoding="utf-8")
df_news3
