==> 2.tokenize_torch.ipynb <==
import pandas as pd
import numpy as np
import torch
device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

print(torch.__version__) 
print(device)
print(torch.cuda.device_count())
print(torch.cuda.get_device_name(0))

import tensorflow as tf
from transformers import BertTokenizer, BertModel, BertForSequenceClassification, AdamW

# 讀取tsv # \t tab做區隔
df_news = pd.read_csv("https://github.com/roccqqck/news_bert/raw/master/data/2015_Company.tsv", sep="\t", encoding="utf-8")
df_news['text'] = df_news['text'].astype(str)
df_news

# Load pre-trained model tokenizer, to convert our text into tokens that correspond to BERT’s vocabulary.
tokenizer = BertTokenizer.from_pretrained('bert-base-chinese')

# Tokenize input
text = "很好看的動作片，不會浪費錢跟時間。很久沒有這樣的探險片。可說是女版的印第安那瓊。"
tokenized_text = tokenizer.tokenize(text)      # 每個字切詞成一個list
print(type(tokenized_text))                 # list
np.array(tokenized_text)                    # 轉成numpy

indexed_tokens = tokenizer.convert_tokens_to_ids(tokenized_text)   # 每個字轉成id
print(type(indexed_tokens))                         # list
np.array(indexed_tokens)

indexed_tokens2 = tokenizer.build_inputs_with_special_tokens(indexed_tokens)    # 句子前後加上 CLS SEP
print(type(indexed_tokens2))
np.array(indexed_tokens2)

n = 512 - len(indexed_tokens2)
indexed_tokens3 = np.pad(indexed_tokens2, (0, n), mode ='constant', constant_values=(0))  
# array右邊append n 個 0  補長度到512
print(len(indexed_tokens3))
indexed_tokens3

segments_ids = tokenizer.create_token_type_ids_from_sequences(indexed_tokens2)
print(type(segments_ids))                                # list
np.array(segments_ids)

#### np.pad doc

https://docs.scipy.org/doc/numpy/reference/generated/numpy.pad.html 

也可以使用
from keras.preprocessing.sequence import pad_sequences

def tokenize_all(text):
#    tokenizer = BertTokenizer.from_pretrained('bert-base-chinese')
    tokenized_text = tokenizer.tokenize(text)    
    indexed_tokens = tokenizer.convert_tokens_to_ids(tokenized_text)
    indexed_tokens2 = tokenizer.build_inputs_with_special_tokens(indexed_tokens)
    indexed_tokens2 = np.array(indexed_tokens2)          # list 轉 numpy
    if len(indexed_tokens2) < 512:
        n = 512 - len(indexed_tokens2)
        indexed_tokens3 = np.pad(indexed_tokens2, (0, n), mode ='constant', constant_values=(0))  # array右邊append n 個 0  補長度到512
    return indexed_tokens3

def segments_all(text):
    tokenized_text = tokenizer.tokenize(text)    
    indexed_tokens = tokenizer.convert_tokens_to_ids(tokenized_text)
    indexed_tokens2 = tokenizer.build_inputs_with_special_tokens(indexed_tokens)
    indexed_tokens2 = np.array(indexed_tokens2)          # list 轉 numpy
    segments_ids = tokenizer.create_token_type_ids_from_sequences(indexed_tokens2)
    segments_ids = np.array(segments_ids)              # list 轉numpy
    if len(segments_ids) < 512:
        n = 512 - len(segments_ids)
        segments_ids2 = np.pad(segments_ids, (0, n), mode ='constant', constant_values=(0))  # array右邊append n 個 0  補長度到512    
    return segments_ids2

def masks_all(text):
    tokenized_text = tokenizer.tokenize(text)    
    indexed_tokens = tokenizer.convert_tokens_to_ids(tokenized_text)
    indexed_tokens2 = tokenizer.build_inputs_with_special_tokens(indexed_tokens)
    indexed_tokens2 = np.array(indexed_tokens2)          # list 轉 numpy
    segments_ids = tokenizer.create_token_type_ids_from_sequences(indexed_tokens2)
    segments_ids = np.array(segments_ids)              # list 轉numpy
    mask_token = np.array([1,1])
    mask_token = np.pad(mask_token, (0, len( segments_ids - 2 )), mode ='constant', constant_values=(1)) 
    # array右邊append 1 到跟segment一樣長
    if len(mask_token) < 512:
        n = 512 - len(mask_token)
        mask_token2 = np.pad(mask_token, (0, n), mode ='constant', constant_values=(0))  # array右邊append n 個 0  補長度到512
    return mask_token2

# df['text'] = df.comment.apply(lambda x: " ".join(jieba.cut(x)))
# df_news['tokenized_text'] = df_news['text'].apply(lambda x:  tokenizer.tokenize(x) )
df_news['indexed_tokens'] = df_news['text'].apply(tokenize_all)
df_news['segments_ids'] = df_news['text'].apply(segments_all)
df_news['mask_token'] = df_news['text'].apply(masks_all)
df_news

print(len(df_news['indexed_tokens'][0]))
df_news['indexed_tokens'][0] 

print(len(df_news['mask_token'][0]))
df_news['mask_token'][0]

print(len(df_news['segments_ids'][0]))
df_news['segments_ids'][0]

# df_news['indexed_tokens'].to_numpy() 出來不是一個2d numpy 
# 只好用for loop一個一個拿出來合併

indexed_tokens = np.zeros((1, 512)).astype(int) #宣吿一個都是0的1*512 numpy # np.zeros預設是float 改成int 不然bert餵不進去
for index, row in df_news.iterrows():  
    element = df_news.loc[index,'indexed_tokens']
    indexed_tokens = np.vstack((indexed_tokens, np.array([element])))   # 2維 合併

indexed_tokens = np.delete(indexed_tokens, 0, 0)              # 刪掉一開始都是0的那一個宣告  
indexed_tokens

indexed_tokens.shape

torch.tensor(indexed_tokens)

tf.convert_to_tensor(indexed_tokens)

mask_token = np.zeros((1, 512)).astype(int) #宣吿一個都是0的1*512 numpy # np.zeros預設是float 改成int 不然bert餵不進去
for index, row in df_news.iterrows():  
    element = df_news.loc[index,'mask_token']
    mask_token = np.vstack((mask_token, np.array([element])))   # 2維 合併

mask_token = np.delete(mask_token, 0, 0)              # 刪掉一開始都是0的那一個宣告  
mask_token

mask_token.shape

label = df_news["label"].to_numpy()
label

https://towardsdatascience.com/https-medium-com-chaturangarajapakshe-text-classification-with-transformer-models-d370944b50ca
https://github.com/ThilinaRajapakse/pytorch-transformers-classification

https://leemeng.tw/attack_on_bert_transfer_learning_in_nlp.html

https://mccormickml.com/2019/07/22/BERT-fine-tuning/

https://blog.csdn.net/Real_Brilliant/article/details/84880528

# 把2個input indexed_tokens , mask_token 還有label 切成training validation

from sklearn.model_selection import train_test_split
# Use train_test_split to split our data into train and validation sets for training

train_inputs, validation_inputs, train_labels, validation_labels = train_test_split(indexed_tokens, label, 
                                                            random_state=2018, test_size=0.1)
train_masks, validation_masks, _, _ = train_test_split(mask_token, indexed_tokens,
                                             random_state=2018, test_size=0.1)

train_inputs = torch.tensor(train_inputs)
validation_inputs = torch.tensor(validation_inputs)
train_labels = torch.tensor(train_labels)
validation_masks = torch.tensor(validation_masks)
train_masks = torch.tensor(train_masks)
validation_labels = torch.tensor(validation_labels)

from torch.utils.data import TensorDataset, DataLoader, RandomSampler
# Select a batch size for training. For fine-tuning BERT on a specific task, the authors recommend a batch size of 16 or 32
batch_size = 32

# Create an iterator of our data with torch DataLoader. This helps save on memory during training because, unlike a for loop, 
# with an iterator the entire dataset does not need to be loaded into memory

train_data = TensorDataset(train_inputs, train_masks, train_labels)
train_sampler = RandomSampler(train_data)
train_dataloader = DataLoader(train_data, sampler=train_sampler, batch_size=batch_size)

# validation_data = TensorDataset(validation_inputs, validation_masks, validation_labels)
# validation_sampler = SequentialSampler(validation_data)
# validation_dataloader = DataLoader(validation_data, sampler=validation_sampler, batch_size=batch_size)


# Load BertForSequenceClassification, the pretrained BERT model with a single linear classification layer on top.
# num_labels=5 分5類
model = BertForSequenceClassification.from_pretrained('bert-base-chinese', num_labels=5)
model.cuda()





# Batch size: 16, 32
# Learning rate (Adam): 5e-5, 3e-5, 2e-5
# Number of epochs: 2, 3, 4
param_optimizer = list(model.named_parameters())
no_decay = ['bias', 'gamma', 'beta']
optimizer_grouped_parameters = [
    {'params': [p for n, p in param_optimizer if not any(nd in n for nd in no_decay)],
     'weight_decay_rate': 0.01},
    {'params': [p for n, p in param_optimizer if any(nd in n for nd in no_decay)],
     'weight_decay_rate': 0.0}
]






# This variable contains all of the hyperparemeter information our training loop needs
optimizer = AdamW(optimizer_grouped_parameters,
                     lr=2e-5,)

outputs = model(train_inputs, train_masks, train_labels)
# loss, logits = outputs[:2]

# Function to calculate the accuracy of our predictions vs labels
def flat_accuracy(preds, labels):
    pred_flat = np.argmax(preds, axis=1).flatten()
    labels_flat = labels.flatten()
    return np.sum(pred_flat == labels_flat) / len(labels_flat)

from tqdm import tqdm, trange

# Store our loss and accuracy for plotting
train_loss_set = []

# Number of training epochs (authors recommend between 2 and 4)
epochs = 4

# trange is a tqdm wrapper around the normal python range
for _ in trange(epochs, desc="Epoch"):
  
  
  # Training
  
  # Set our model to training mode (as opposed to evaluation mode)
  model.train()
  
  # Tracking variables
  tr_loss = 0
  nb_tr_examples, nb_tr_steps = 0, 0
  
  # Train the data for one epoch
  for step, batch in enumerate(train_dataloader):
    # Add batch to GPU
    batch = tuple(t.to(device) for t in batch)
    # Unpack the inputs from our dataloader
    b_input_ids, b_input_mask, b_labels = batch
    # Clear out the gradients (by default they accumulate)
    optimizer.zero_grad()
    # Forward pass
    loss = model(b_input_ids, token_type_ids=None, attention_mask=b_input_mask, labels=b_labels)
    train_loss_set.append(loss.item())    
    # Backward pass
    loss.backward()
    # Update parameters and take a step using the computed gradient
    optimizer.step()
    
    
    # Update tracking variables
    tr_loss += loss.item()
    nb_tr_examples += b_input_ids.size(0)
    nb_tr_steps += 1

  print("Train loss: {}".format(tr_loss/nb_tr_steps))
    
    
  # Validation

  # Put model in evaluation mode to evaluate loss on the validation set
  model.eval()

  # Tracking variables 
  eval_loss, eval_accuracy = 0, 0
  nb_eval_steps, nb_eval_examples = 0, 0

  # Evaluate data for one epoch
  for batch in validation_dataloader:
    # Add batch to GPU
    batch = tuple(t.to(device) for t in batch)
    # Unpack the inputs from our dataloader
    b_input_ids, b_input_mask, b_labels = batch
    # Telling the model not to compute or store gradients, saving memory and speeding up validation
    with torch.no_grad():
          # Forward pass, calculate logit predictions
          logits = model(b_input_ids, token_type_ids=None, attention_mask=b_input_mask)
    
    # Move logits and labels to CPU
    logits = logits.detach().cpu().numpy()
    label_ids = b_labels.to('cpu').numpy()

    tmp_eval_accuracy = flat_accuracy(logits, label_ids)
    
    eval_accuracy += tmp_eval_accuracy
    nb_eval_steps += 1

  print("Validation Accuracy: {}".format(eval_accuracy/nb_eval_steps))

input_ids = torch.tensor(tokenizer.encode("很好看的動作片，不會浪費錢跟時間。很久沒有這樣的探險片。可說是女版的印第安那瓊。")).unsqueeze(0)  # Batch size 1
input_ids

labels = torch.tensor([1]).unsqueeze(0)  # Batch size 1
labels

outputs = model(input_ids, labels=labels)
loss, logits = outputs[:2]

outputs

loss, logits
