==> 6.news_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np
import re
import os

from tqdm import tqdm
tqdm.pandas()

# 1 匯入資料

## 1-1 匯入危機公司新聞

### 1-1-1 匯入TEJ危機公司新聞

tej_distress = pd.read_csv('data/2018_2017_tej_distress_0423.csv')
print(tej_distress.shape)
tej_distress.head(2)

### 1-1-2 匯入UDN危機公司新聞

# 設定路徑，將該資料夾內所有檔案名稱列出來
Folder_Path = 'data/udn_distress/'
file_list = os.listdir(Folder_Path)
# file_list

udn_distress = pd.DataFrame(columns=['company', 'title', 'time', 'content'])

for file in file_list:
    df_temp = pd.read_csv('data/udn_distress/' + file)
    udn_distress = pd.concat([udn_distress, df_temp])
udn_distress = udn_distress.drop('title', axis=1).dropna()
print(udn_distress.shape)
udn_distress.head(2)

### 1-1-3 （針對UDN）擷取日期，把日期格式轉換為yyyy/mm/dd

from datetime import date
import datetime

#測試一下

s = udn_distress['time'].iloc[4]
print(s)
match = re.search('\d+-\d+-\d+', s)
date = str(match.group(0))
date = date.replace('-', '/')
print(date)
date = datetime.datetime.strptime(date, '%Y/%m/%d').date()
# date = datetime.datetime.strptime(match.group(0), '%Y-%m-%d').date()
print(date)

dates = []
for time in udn_distress['time']:
    s = time
    match = re.search('\d+-\d+-\d+', s)
    date = str(match.group(0))
    date = date.replace('-', '/')
    dates.append(date)
udn_distress['date'] = dates
# udn_distress.drop('time', axis=1, inplace=True)
print(udn_distress.shape)
udn_distress.head(2)

# 把一些字串刪掉改空格
udn_distress['content'] = udn_distress['content'].astype(str).str.replace('\r' , ' ').str.replace('\t' , ' ').str.replace("\n", " ").str.replace(",", " ").str.replace('"' , ' ').str.replace("'" , ' ')
udn_distress['company'] = udn_distress['company'].astype(str).str.replace('\r' , ' ').str.replace('\t' , ' ').str.replace("\n", " ").str.replace(",", " ").str.replace('"' , ' ').str.replace("'" , ' ')
print(udn_distress.shape)
udn_distress.head(2)

udn_distress = udn_distress.loc[:, ['company', 'content', 'date']]
print(udn_distress.shape)
udn_distress.head(2)

udn_distress.to_csv("data/udn_distress_news.csv", index=False)

udn_distress = pd.read_csv("data/udn_distress_news.csv", encoding="utf-8")
print(udn_distress.shape)
udn_distress.head(2)

### 1-1-4 合併危機公司資料集為 distress_news

distress_news = pd.DataFrame(columns=['company', 'date', 'content'])
company = udn_distress['company']
company = company.append(tej_distress['公司碼'])
date = udn_distress['date']
date = date.append(tej_distress['年月'])
content = udn_distress['content']
content = content.append(tej_distress['新聞內容'])
distress_news['company'] = company
distress_news['date'] = date
distress_news['content'] = content
print(distress_news.shape)
distress_news.head(2)

distress_news.to_csv("data/tej_udn_distress_news.csv", index=False)

distress_news = pd.read_csv("data/tej_udn_distress_news.csv", encoding="utf-8")
print(distress_news.shape)
distress_news.head(2)

## 1-2 匯入正常公司新聞

### 1-2-1 匯入TFJ正常公司新聞

tej_normal = pd.read_csv('data/2018_2017_tej_normal_0423.csv')
print(tej_normal.shape)
tej_normal.head(2)

### 1-2-2 匯入UDN正常公司新聞（0422：正常公司的UDN資料蒐集不完整，暫不採納）

# Folder_Path = '/home/michael/Documents/M10609101_Thesis/News/udn/normal/'
# file_list = os.listdir(Folder_Path)
# # file_list

# udn_normal = pd.DataFrame(columns=['company', 'title', 'time', 'content'])

# for file in file_list:
#     df_temp = pd.read_csv('News/udn/normal/' + file)
#     udn_normal = pd.concat([udn_normal, df_temp])
# udn_normal = udn_normal.drop('title', axis=1).dropna()
# udn_normal.head()

# len(udn_normal)

### 1-2-3 （針對UDN）擷取日期，把日期格式轉換為yyyy/mm/dd（0422：正常公司的UDN資料蒐集不完整，暫不採納）

# dates = []
# for time in udn_normal['time']:
#     s = time
#     match = re.search('\d+-\d+-\d+', s)
#     date = str(match.group(0))
#     date = date.replace('-', '/')
#     dates.append(date)
# udn_normal['date'] = dates
# # udn_normal.drop('time', axis=1, inplace=True)
# udn_normal.tail()

### 1-2-4 合併危機公司資料集為 normal_news

normal_news = pd.DataFrame(columns=['company', 'date', 'content'])
normal_news['company'] = tej_normal['公司碼']
normal_news['date'] = tej_normal['年月']
normal_news['content'] = tej_normal['新聞內容']
print(normal_news.shape)
normal_news.head(2)

# normal_news = pd.DataFrame(columns=['company', 'date', 'content'])
# company = udn_normal['company']
# company = company.append(tej_normal['公司碼'])
# date = udn_normal['date']
# date = date.append(tej_normal['年月'])
# content = udn_normal['content']
# content = content.append(tej_normal['新聞內容'])
# normal_news['company'] = company
# normal_news['date'] = date
# normal_news['content'] = content
# normal_news.head()

normal_news.to_csv("data/tej_normal_news.csv", index=False)

normal_news = pd.read_csv("data/tej_normal_news.csv", encoding="utf-8")
print(normal_news.shape)
normal_news.head(2)

all_news = pd.concat([normal_news, distress_news])   # 上下合併
print(all_news.shape)
all_news.head(2)

all_news['date'] = pd.to_datetime(all_news['date'])
all_news = all_news.sort_values(by=["date","company"]).reset_index(drop=True) # 用post_time排序 # 在重設index
print(all_news.shape)
all_news.head(2)

all_news.to_csv("data/tej_udn_all_news.csv", index=False)

## 2-3 每間公司風險分布
使用意藍2015年風險資料集、經過BERT主題模型產生的model 將每篇文章輸出1*5的output

13302篇文章就有13302*5

import pandas as pd
import numpy as np
import re
import os

from tqdm import tqdm
tqdm.pandas()

import tensorflow as tf
from transformers import BertTokenizer, TFBertModel, TFBertForSequenceClassification, AdamW

from tensorflow.keras import Sequential, Model, Input
from tensorflow.keras.layers import Dense, Embedding, SpatialDropout1D, Dropout, Activation, Flatten, InputLayer

all_news = pd.read_csv("data/tej_udn_all_news.csv", encoding="utf-8")
all_news = all_news.sort_values(by=["date","company"]).reset_index(drop=True)
print(all_news.shape)
all_news.head(2)

all_news.describe()

all_news.info()

# 文章字數 > 510了話 去尾
# 字數小於512-2 因為還有CLS SEP
def remove_510(text):
    if len(text) > 510:
        text = text[:510]    # 只取前510個字
    return text

all_news["content"] = all_news["content"].progress_apply(remove_510)
print(all_news.shape)
all_news.head(2)

# Load pre-trained model tokenizer, to convert our text into tokens that correspond to BERT’s vocabulary.
tokenizer = BertTokenizer.from_pretrained('bert-base-chinese')

def input_ids_all(text):
#    tokenizer = BertTokenizer.from_pretrained('bert-base-chinese')
    tokens = tokenizer.tokenize(text)        # 每個字切詞成一個list
    input_ids = tokenizer.convert_tokens_to_ids(tokens)  # 每個字轉成id
    input_ids = tokenizer.build_inputs_with_special_tokens(input_ids)    # 句子前後加上 CLS SEP 的 id
    input_ids = np.array(input_ids)          # list 轉 numpy
    if len(input_ids) < 512:
        n = 512 - len(input_ids)
        input_ids = np.pad(input_ids, (0, n), mode ='constant', constant_values=(0))  
        # array右邊append n 個 0  補長度到512
    return input_ids

# df['text2'] = df['text']].apply(lambda x: " ".join(jieba.cut(x)))
# df_news['tokens'] = df_news['text'].apply(lambda x:  tokenizer.tokenize(x) )
all_news['input_ids'] = all_news['content'].progress_apply(input_ids_all)
print(all_news.shape)
all_news.head(2)

# df_news['input_ids'].to_numpy() 出來不是一個2d numpy 
# 只好用for loop一個一個拿出來合併
# 用np vstack超級慢 不知道為何  改用最外層是list append
input_ids = []      # list
for index, row in all_news.iterrows():  
    np_1d = all_news.loc[index,'input_ids']    # 1d np arrary
    input_ids.append(np_1d)                       # 1d np的 list # list[np_1, np_2, np_3, ....]


input_ids = np.array(input_ids)                       # 轉成2d np
print(input_ids.shape)
input_ids

# Recreate the exact same model, including its weights and the optimizer
# 讀取存的model
input_layer = Input(shape = (512,), dtype='int64')  # 預設是float 要改成input_id的int64 # tensorflow2.0 bug 修正後可能不需要這行
load_model = tf.keras.models.load_model('model/model_bert_eland_softmax_1')(input_layer)
model_1 = Model(inputs=input_layer, outputs=load_model) # tensorflow2.0 bug 修正後可能不需要這行

# Show the model architecture
model_1.summary()

%%time
predictions = model_1.predict(input_ids)   # 輸出的是n*5的編碼值array
print(predictions.shape)
predictions

df_predictions = pd.DataFrame(predictions, columns = ["c_0", "c_1", "c_2", "c_3", "c_4"])
print(df_predictions.shape)
df_predictions.head(2)

all_news = pd.concat([all_news, df_predictions], axis=1)   # axis=1左右合併 axis=0 上下合併
print(all_news.shape)
all_news.head(2)

all_news["c_0"].sum()

### 需要宣告一個240公司*10年的dataframe

company_list = all_news.sort_values(by=["company","date"])["company"].unique()  # 把欄位unique的值都挑出來變成numpy
print(len(company_list))
company_list

temp = []
for company_name in tqdm(company_list):
    for year in range(2008,2018):
        company_name = str(company_name)
        year = int(year)
        temp_np = np.array([company_name , year])    # np(公司名,年分)
        temp.append(temp_np)
        
temp = np.array(temp)
print(len(temp))
temp

df_company_year_prob_sum = pd.DataFrame(temp, columns = ['company' , 'date'])  # numpy to pandas 並宣告欄位名
print(df_company_year_prob_sum.shape)
df_company_year_prob_sum

df_company_year_prob_sum["date"] = df_company_year_prob_sum["date"].astype(int)
df_company_year_prob_sum["company"] = df_company_year_prob_sum["company"].astype(str)
df_company_year_prob_sum["c_0"] = 0.       # 建class_0 ~ class_4的欄位
df_company_year_prob_sum["c_1"] = 0.
df_company_year_prob_sum["c_2"] = 0.
df_company_year_prob_sum["c_3"] = 0.
df_company_year_prob_sum["c_4"] = 0.
print(df_company_year_prob_sum.shape)
df_company_year_prob_sum.head(2)

# pandas 取出 a column contains "A" & b column contains "B" 
# df_company_year_prob_sum[(df_company_year_prob_sum["company"].str.contains("楠梓電")) & (df_company_year_prob_sum["date"].astype(str).str.contains("2009") )]

df_company_year_prob_sum.loc[(df_company_year_prob_sum["company"].str.contains("楠梓電")) 
                    & (df_company_year_prob_sum["date"].astype(str).str.contains("2009") )]

### 把每公司 每年 的 class_0相加
### 再丟到df_topic_5_prob

# pandas all_news 取出 company column contains "楠梓電" & date column contains "2008" 
# df_temp = all_news[(all_news["company"].str.contains("楠梓電")) & (all_news["date"].astype(str).str.contains("2008") )]

df_temp = all_news.loc[(all_news["company"].str.contains("楠梓電"))
                    & (all_news["date"].astype(str).str.contains("2008"))]
print(df_temp.shape)
df_temp.head()

# 將 c_0欄位 加總sum()
df_temp["c_0"].sum()

# df_company_year_prob_sum[(df_company_year_prob_sum["company"].str.contains("楠梓電") ) & (df_company_year_prob_sum["date"] == 2008) ]["c_0"] = df_temp["c_0"].sum()
# 填入 company欄位 contains 楠梓電 date欄位 contains 2008 的 c_0欄位
# 沒df.loc 填入表格會失敗

df_company_year_prob_sum.loc[(df_company_year_prob_sum["company"].str.contains("楠梓電") )
                    & (df_company_year_prob_sum["date"].astype(str).str.contains( str(2008) )), "c_0"] = df_temp["c_0"].sum()
df_company_year_prob_sum

# 上面步驟寫成for loop
# for 240公司名list:
#     for 2008-2017年:

for company_name in tqdm(company_list):
    for year in range(2008,2018):
        company_name = str(company_name)
        
        # pandas all_news 取出 company欄位 contains 公司名 & date欄位 contains 年分 
        df_temp = all_news.loc[(all_news["company"].str.contains(company_name))
                    & (all_news["date"].astype(str).str.contains( str(year) ))]
        # 將 c_0 c_1 c_2 c_3 c_4 欄位 加總sum()
        # df_temp["c_0"].sum(), df_temp["c_1"].sum(), df_temp["c_2"].sum() ,df_temp["c_3"].sum() ,df_temp["c_4"].sum()

        # 將 c_0 ~ c_4 欄位 加總sum()
        # 按照 公司名 年分 填入 df_company_year_prob_sum的 c_0 ~ c_4 欄位表格
        df_company_year_prob_sum.loc[(df_company_year_prob_sum["company"].str.contains(company_name) )
                    & (df_company_year_prob_sum["date"].astype(str).str.contains( str(year) )), "c_0"] = df_temp["c_0"].sum()
        df_company_year_prob_sum.loc[(df_topic_5_prob["company"].str.contains(company_name) )
                    & (df_company_year_prob_sum["date"].astype(str).str.contains( str(year) )), "c_1"] = df_temp["c_1"].sum()
        df_company_year_prob_sum.loc[(df_company_year_prob_sum["company"].str.contains(company_name) )
                    & (df_company_year_prob_sum["date"].astype(str).str.contains( str(year) )), "c_2"] = df_temp["c_2"].sum()
        df_company_year_prob_sum.loc[(df_company_year_prob_sum["company"].str.contains(company_name) )
                    & (df_company_year_prob_sum["date"].astype(str).str.contains( str(year) )), "c_3"] = df_temp["c_3"].sum()
        df_company_year_prob_sum.loc[(df_company_year_prob_sum["company"].str.contains(company_name) )
                    & (df_company_year_prob_sum["date"].astype(str).str.contains( str(year) )), "c_4"] = df_temp["c_4"].sum()

df_company_year_prob_sum

df_company_year_prob_sum.to_csv("data/company_year_prob_sum.csv", index=False)

df_company_year_prob_sum = pd.read_csv("data/company_year_prob_sum.csv", encoding="utf-8")
df_company_year_prob_sum["c_0"] = df_company_year_prob_sum["c_0"].astype(float)  # 讀取後將欄位弄成float
df_company_year_prob_sum["c_1"] = df_company_year_prob_sum["c_1"].astype(float)
df_company_year_prob_sum["c_2"] = df_company_year_prob_sum["c_2"].astype(float)
df_company_year_prob_sum["c_3"] = df_company_year_prob_sum["c_3"].astype(float)
df_company_year_prob_sum["c_4"] = df_company_year_prob_sum["c_4"].astype(float)
print(df_company_year_prob_sum.shape)
df_company_year_prob_sum

df_company_year_prob_sum.sort_values(by=["date","company"])  # 先日期排序 再公司名排序

df_company_year_prob_sum.sort_values(by=["company","date"])  # 先公司名排序 再日期排序

company_list = df_company_year_prob_sum["company"].unique()  # 把欄位unique的值都挑出來變成numpy
print(len(company_list))
company_list
