==> 7.TCRI_predict.ipynb <==
# import pandas as pd
# import numpy as np
# from sklearn import preprocessing
# from sklearn.model_selection import train_test_split
# from sklearn.svm import SVC
# from sklearn.linear_model import LogisticRegression
# from keras.models import Sequential
# from keras.layers import Dense, Dropout
# from keras.optimizers import SGD, RMSprop, Adam
# from keras.models import load_model

# import tensorflow as tf
# from neupy import algorithms
# from sklearn.metrics import mean_squared_error, confusion_matrix
# from sklearn.metrics import multilabel_confusion_matrix
# from math import sqrt
# from sklearn.tree import DecisionTreeClassifier
# from sklearn import metrics

# # 用隨機森林做Feature Selection
# from sklearn.ensemble import ExtraTreesClassifier
# from sklearn.feature_selection import SelectFromModel

# # ROC Curve
# from scipy import interp
# from itertools import cycle
# from sklearn.metrics import roc_curve, auc

# import re
# from datetime import date
# import datetime

# import matplotlib.pyplot as plt
# import seaborn as sns
# %matplotlib inline

import pandas as pd
import numpy as np
from tqdm import tqdm
tqdm.pandas()

# 1 匯入資料＆前處理

df = pd.read_csv('data/TCRI_2008_2017.csv', encoding="utf-8")
df = df.sort_values(by=["年月","公司"])   # 先日期排序 再公司名排序
df = df.reset_index(drop=True)           # index重新排序
print(df.shape)
df.head(2)

df.describe()

df.info()

## ※匯入TSE產業別代碼

tse = pd.read_csv('data/TSE_category.csv', encoding="utf-8")
print(tse.shape)
tse.head(2)

tse.set_index('公司簡稱').to_dict()
tse_dict = tse.set_index('公司簡稱')['TSE 產業別'].to_dict()

tse_no = []
for company in df['公司']:
    tse_no.append(tse_dict[company])
df['TSE產業別'] = tse_no
df.head(2)

df.to_csv('data/TCRI_2008_2017_tse.csv', index=False)

## 1-1 調整「CFO不一致」欄位

import pandas as pd
import numpy as np
from tqdm import tqdm
tqdm.pandas()

df = pd.read_csv('data/TCRI_2008_2017_tse.csv', encoding="utf-8")
df = df.sort_values(by=["年月","公司"])   # 先日期排序 再公司名排序
df = df.reset_index(drop=True)           # index重新排序
print(df.shape)
df.head(2)

# 將「CFO不一致」欄位的值調整，空值(Nan)與minus sign(-)改為0，plus sign(+)改為1
df['CFO不一致'] = df['CFO不一致'].fillna(0)
df['CFO不一致'] = df['CFO不一致'].replace('+', 1)
df['CFO不一致'] = df['CFO不一致'].replace('-', 0)

## 1-2 調整「TCRI」欄位

# 將TCRI欄位中"-"資料刪掉；D表示違約，另外處理（先暫時改成0）
index_to_drop = df[df['TCRI'] == '-'].index     # 把 df['TCRI'] == '-' 的index 挑出來
df = df.drop(index=index_to_drop, axis=1)       # 刪除指定index 整個row
df['TCRI'] = df['TCRI'].replace('D', 0)        # D表示違約，另外處理（先暫時改成0）

# TCRI原本有非數值的資料，因此原始的型態並非float64，因此需要轉型
df['TCRI'] = df['TCRI'].astype(float)
print(df.shape)
df.head(2)

## 1-3 去除空值與非數值型資料

# 有 空值 NaN 的 row橫列 都刪掉
df.dropna(axis=0, inplace=True)
print(df.shape)

## 1-4 去除「規模」、「獲利」、「安全」、「效率」等4個欄位

df = df.drop(['規模', '獲利', '安全', '效率'], axis=1)
print(df.shape)
df = df.reset_index(drop=True)           # index重新排序
df.tail(2)

## ※合併輿情風險分數

news_score = pd.read_csv("data/company_year_prob_sum.csv", encoding="utf-8")
news_score["c_0"] = news_score["c_0"].astype(float)  # 讀取後將欄位弄成float
news_score["c_1"] = news_score["c_1"].astype(float)
news_score["c_2"] = news_score["c_2"].astype(float)
news_score["c_3"] = news_score["c_3"].astype(float)
news_score["c_4"] = news_score["c_4"].astype(float)
print(news_score.shape)
news_score.tail(2)

# 把 c_0~c_4都等於0 的index 挑出來
index_to_drop = news_score[(news_score['c_0'] == 0.0) &
                           (news_score['c_1'] == 0.0) &
                           (news_score['c_2'] == 0.0) &
                           (news_score['c_3'] == 0.0) &
                           (news_score['c_4'] == 0.0)].index     # 把 c_0~c_4都等於0 的 row的 index 挑出來
print(len(index_to_drop))
index_to_drop

news_score = news_score.drop(index=index_to_drop, axis=1)       # 刪除指定index 整個row
news_score = news_score.reset_index(drop=True)           # index重新排序
print(news_score.shape)
news_score.tail(2)

#### 因為df 的row數 跟 news_score不一樣
#### 比對 company和year 再把 news_score表格的欄位 填入 df 表格

# # M106陳柏霖寫法

# risk1 = []
# risk2 = []
# risk3 = []
# risk4 = []
# risk5 = []
# risk6 = []
# # 年份跟公司也append，方便確認每家公司的風險分數沒有對應錯誤
# year_from_news_score = []
# company_from_news_score = []

# for i in range(len(df)):
#     s = df['年月'].iloc[i]
#     match = re.search('\d+/\d+/\d+', s)
#     date = datetime.datetime.strptime(match.group(0), '%Y/%m/%d').date()
#     year = date.year
#     company = df['公司'].iloc[i]
#     isInsert = False
#     for j in range(len(news_score)):
#         if (year == news_score['Year'].iloc[j]) and (company == news_score['company'].iloc[j]):
#             risk1.append(np.float64(news_score['0'].iloc[j]))
#             risk2.append(np.float64(news_score['1'].iloc[j]))
#             risk3.append(np.float64(news_score['2'].iloc[j]))
#             risk4.append(np.float64(news_score['3'].iloc[j]))
#             risk5.append(np.float64(news_score['4'].iloc[j]))
#             risk6.append(np.float64(news_score['5'].iloc[j]))
#             year_from_news_score.append(news_score['Year'].iloc[j])
#             company_from_news_score.append(news_score['company'].iloc[j])
#             isInsert = True
#     if not isInsert:
#         risk1.append(0.0)
#         risk2.append(0.0)
#         risk3.append(0.0)
#         risk4.append(0.0)
#         risk5.append(0.0)
#         risk6.append(0.0)
#         year_from_news_score.append(0)
#         company_from_news_score.append('NO')

# # M106陳柏霖寫法

# df['財務風險'] = risk1
# df['管理風險'] = risk2
# df['法律風險'] = risk3
# df['供應鏈風險'] = risk4
# df['誠信風險'] = risk5
# df['勞動風險'] = risk6
# df['year'] = year_from_news_score
# df['company'] = company_from_news_score
# df.tail()

df["c_0"] = 0.0
df["c_1"] = 0.0
df["c_2"] = 0.0
df["c_3"] = 0.0
df["c_4"] = 0.0
df.tail(2)

company_list = news_score["company"].astype(str).unique()  # 把欄位unique的值都挑出來變成numpy
print(len(company_list))
company_list[:10]

# news_score表格 取出"company" contains"三發地產" 並且 "date" contains 2017的表格
df_temp = news_score.loc[(news_score["company"].str.contains("三發地產"))
                          & (news_score["date"].astype(str).str.contains( str(2017) ))]
df_temp

# df_temp c_0欄位取值 用sum是因為有時候會遇到df_temp表格是空的
df_temp["c_0"].sum()  

# df表格 取出"company" contains"三發地產" 並且 "date" contains 2017的表格
df.loc[(df["公司"].astype(str).str.contains("三發地產") )
                      & (df["年月"].astype(str).str.contains( str(2017) ))]

# df表格 取出"company" contains"三發地產" 並且 "date" contains 2017的表格 的 c_0欄位 = df_temp["c_0"].sum()
df.loc[(df["公司"].astype(str).str.contains("三發地產") )
                      & (df["年月"].astype(str).str.contains( str(2017) )), "c_0"] = df_temp["c_0"].sum()
df.tail(2)

# # 用pandas兩個欄位的for loop應該比較快 但我懶惰重新寫
# for company_name, year in tqdm(news_score[["company", "date"]].to_numpy(), total=428):
#     print(company_name, year)

# 把上面寫成for loop
# for 240公司名list:
#     for 2008-2017年:

for company_name in tqdm(company_list):
    for year in range(2008,2018):
        company_name = str(company_name)
        
        # news_score 取出 company欄位 contains 公司名 & date欄位 contains 年分 
        df_temp = news_score.loc[(news_score["company"].str.contains(company_name))
                       & (news_score["date"].astype(str).str.contains( str(year) ))]


        # 將 c_0 ~ c_4 欄位 
        # 按照 公司名 年分 填入 df 的 c_0 ~ c_4 欄位表格
        df.loc[(df["公司"].astype(str).str.contains(company_name) )
                    & (df["年月"].astype(str).str.contains( str(year) )), "c_0"] = df_temp["c_0"].sum()
        df.loc[(df["公司"].astype(str).str.contains(company_name) )
                    & (df["年月"].astype(str).str.contains( str(year) )), "c_1"] = df_temp["c_1"].sum()
        df.loc[(df["公司"].astype(str).str.contains(company_name) )
                    & (df["年月"].astype(str).str.contains( str(year) )), "c_2"] = df_temp["c_2"].sum()
        df.loc[(df["公司"].astype(str).str.contains(company_name) )
                    & (df["年月"].astype(str).str.contains( str(year) )), "c_3"] = df_temp["c_3"].sum()
        df.loc[(df["公司"].astype(str).str.contains(company_name) )
                    & (df["年月"].astype(str).str.contains( str(year) )), "c_4"] = df_temp["c_4"].sum()

df.tail(2)

### 因為有些年有些公司沒有文章 看機率全為0.0的有幾筆

# 把 c_0~c_4都等於0 的index 挑出來
index_to_drop = df[(df['c_0'] == 0.0) &
                   (df['c_1'] == 0.0) &
                   (df['c_2'] == 0.0) &
                   (df['c_3'] == 0.0) &
                   (df['c_4'] == 0.0)].index     # 把 c_0~c_4都等於0 的 row的 index 挑出來
print(len(index_to_drop))
index_to_drop

df2 = df.drop(index=index_to_drop, axis=1)       # 刪除指定index 整個row
df2 = df2.reset_index(drop=True)           # index重新排序
print(df2.shape)
del df2

df.to_csv('data/TCRI_2008_2017_prob.csv', index=False)

# 2 Feature Engineering

## 2-1 根據基本總分產生的初步TCRI等級

import pandas as pd
import numpy as np
from tqdm import tqdm
tqdm.pandas()

df = pd.read_csv("data/TCRI_2008_2017_prob.csv", encoding="utf-8")
df["c_0"] = df["c_0"].astype(float)  # 讀取後將欄位弄成float
df["c_1"] = df["c_1"].astype(float)
df["c_2"] = df["c_2"].astype(float)
df["c_3"] = df["c_3"].astype(float)
df["c_4"] = df["c_4"].astype(float)
df["TCRI"] = df["TCRI"].astype(float)
print(df.shape)
df.tail(2)

def transformBasicTCRILevel(value):
    level = 0
    if value < 165.0:
        level = 9
    elif value < 250.0:
        level = 8
    elif value < 335.0:
        level = 7
    elif value < 420.0:
        level = 6
    elif value < 505.0:
        level = 5
    elif value < 590.0:
        level = 4
    elif value < 675.0:
        level = 3
    elif value < 760.0:
        level = 2
    else:
        level = 1
    return level

basicTCRI = []
for score in df['總分']:
    basicTCRI.append(transformBasicTCRILevel(score))

from sklearn.metrics import mean_squared_error
from math import sqrt
rmse = sqrt(mean_squared_error(basicTCRI, df['TCRI']))
print(rmse)

# 用初步等級將總分替換
df.drop(['總分'], axis=1, inplace=True)
df['初步等級'] = basicTCRI
df.head()

## 2-2 門檻等級：規模

thresholdLevel = []
for i in range(len(df)):
    rev = df['營收'].iloc[i]
    level = df['初步等級'].iloc[i]
    threshold = 0
    
    # 營收不到5億，則不列入前5等；電子業門檻較高，門檻為10億
    if df['TSE產業別'].iloc[i] == 28: # 電子零組件業產業別代碼為28
        threshold = 10
    else:
        threshold = 5
    if rev < threshold and level <= 5:
        level = 6
        
    # 營收不到10億，則不列入前4等；電子業門檻較高，門檻為20億
    if df['TSE產業別'].iloc[i] == 28: # 電子零組件業產業別代碼為28
        threshold = 20
    else:
        threshold = 10
    if rev < threshold and level <= 4:
        level = 5
    
    # 營收不到30億，則不列入前3等；電子業門檻較高，門檻為40億
    if df['TSE產業別'].iloc[i] == 28: # 電子零組件業產業別代碼為28
        threshold = 40
    else:
        threshold = 30
    if rev < threshold and level <= 3:
        level = 4
        
    # 營收不到100億，則不列入前2等
    threshold = 100
    if rev < threshold and level <= 2:
        level = 3
        
    # 營收不到200億，則不列入前1等
    threshold = 200
    if rev < threshold and level <= 1:
        level = 2
    
    # 補充電子業門檻較高原因：受景氣影響較深，除非擁有關鍵技術或特殊的利基市場、獨門生意，否則一律適用較嚴格的規模門檻。
    
    thresholdLevel.append(level)

df['門檻等級_規模'] = thresholdLevel
df.head()

## 2-2 APL 財務壓力計

### 2-2-1 Assets Management

# # 確認有無離群值
# plt.subplot(1, 2, 1)
# plt.boxplot(df['平均收款天增減'])
# plt.subplot(1, 2, 2)
# plt.boxplot(df['平均售貨天增減'])
# plt.tight_layout()

from scipy import stats
index_temp_1 = df[(np.abs(stats.zscore(df['平均收款天增減'])) > 3)].index
index_temp_2 = df[(np.abs(stats.zscore(df['平均售貨天增減'])) > 3)].index
# display(index_temp_1)
# display(index_temp_2)

# 聯集：讓index不重複
def find_union(lst1, lst2):
    final_list = list(set(lst1) | set(lst2)) 
    return final_list

index_to_drop = find_union(index_temp_1, index_temp_2)

df.drop(index=index_to_drop, axis=1, inplace=True)
df.head(2)

# plt.subplot(1, 2, 1)
# plt.boxplot(df['平均收款天增減'])
# plt.subplot(1, 2, 2)
# plt.boxplot(df['平均售貨天增減'])
# plt.tight_layout()

# A1. 平均收帳天數比前一期慢 >= 20 天
# df[df['平均收款天增減'] >= 20]
A1 = []
for d in df['平均收款天增減']:
    if d >= 20:
        A1.append(int(d / 20))
    else:
        A1.append(0)
df['A1'] = A1

# A3. 平均售貨天數比前一期慢 >= 20 天
A3 = []
for d in df['平均售貨天增減']:
    if d >= 20:
        A3.append(int(d / 20))
    else:
        A3.append(0)
df['A3'] = A3

### 2-2-2 Profitability

# P1. 當期的總資產報酬率(ROA)不敷借款息
# df[df['資產報酬%'] - df['借款息%'] <= 0]
P1 = []
for i in range(len(df)):
    if df['資產報酬%'].iloc[i] - df['借款息%'].iloc[i] <= 0:
        P1.append(True)
    else:
        P1.append(False)

# P2. 當期的營業利益率(OP%)不足支應利息支出率(INT%)
# df[df['營業利益%'] - df['利息支出率'] <= 0]
P2 = []
for i in range(len(df)):
    if df['營業利益%'].iloc[i] - df['利息支出率'].iloc[i] <= 0:
        P2.append(True)
    else:
        P2.append(False)

'''
P發生的case
(1) P1且P2
(2) 連續3期P1（先略過）
'''
P_case1 = []
for i in range(len(df)):
    if P1[i] and P2[i]:
        P_case1.append(1)
    else:
        P_case1.append(0)
df['P'] = P_case1
df.head()

### 2-2-3 Liquidity

# L1. 借款依存度(DE%)超過100%
# df[df['借款依存度%'] > 100]
L1 = []
for l in df['借款依存度%']:
    if l > 100:
        L1.append(1)
    else:
        L1.append(0)
df['L1'] = L1

# L3. CFO為負，不足資金 < -0.25億
# df[df['CFO'] < -0.25]
L3 = []
for l in df['CFO']:
    if l < 0.25:
        L3.append(1)
    else:
        L3.append(0)
df['L3'] = L3

df.head()

# 3 訓練前的準備

## 3-1 Resample imbalanced labels
https://www.kaggle.com/rafjaa/resampling-strategies-for-imbalanced-datasets

df['TCRI'].value_counts()

# df['TCRI'].value_counts().plot(kind='bar')

# Class count
count_TCRI_4 = 2*len(df[df['TCRI'] == 4])


# Divide by class
df_TCRI_0 = df[df['TCRI'] == 0.0]
df_TCRI_1 = df[df['TCRI'] == 1]
df_TCRI_2 = df[df['TCRI'] == 2]
df_TCRI_3 = df[df['TCRI'] == 3]
df_TCRI_4 = df[df['TCRI'] == 4]
df_TCRI_5 = df[df['TCRI'] == 5]
df_TCRI_6 = df[df['TCRI'] == 6]
df_TCRI_7 = df[df['TCRI'] == 7]
df_TCRI_8 = df[df['TCRI'] == 8]
df_TCRI_9 = df[df['TCRI'] == 9]

# 把每個類別都複製到跟4類別一樣多

df_TCRI_0_over = df_TCRI_0.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_1_over = df_TCRI_1.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_2_over = df_TCRI_2.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_3_over = df_TCRI_3.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_4_over = df_TCRI_4.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_5_over = df_TCRI_5.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_6_over = df_TCRI_6.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_7_over = df_TCRI_7.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_8_over = df_TCRI_8.sample(n=count_TCRI_4, replace=True, random_state=2018, )
df_TCRI_9_over = df_TCRI_9.sample(n=count_TCRI_4, replace=True, random_state=2018, )

# 新的複製的data 合併
df_resample = pd.concat([df_TCRI_0_over, df_TCRI_1_over, df_TCRI_2_over, df_TCRI_3_over, df_TCRI_4_over, df_TCRI_5_over, df_TCRI_6_over, df_TCRI_7_over, df_TCRI_8_over, df_TCRI_9_over])
df_resample['TCRI'].value_counts()

# df_resample['TCRI'].value_counts().plot(kind='bar')

## 3-2 資料前處理 function

def dataPreprocessing(df, IsSelected, HasCrisis, GenMoreData):
    # 3-1 區分Feature (X) 跟Label (Y)
    Y = df['TCRI'].values
    
    if HasCrisis:
        X = df.drop(['TCRI', '公司', '年月'], axis = 1).values
    else:
        X = df.drop(['TCRI', '公司', '年月', 'c_0', 'c_1', 'c_2', 'c_3', 'c_4' ], axis = 1).values
    
    if IsSelected: # Feature Selection with Random Forest
        clf = ExtraTreesClassifier(n_estimators=50)
        clf = clf.fit(X, Y)
        model = SelectFromModel(clf, prefit=True)
        X = model.transform(X)
        X.shape
    
    # 3-2 標準化
#     minmax_scale = preprocessing.MinMaxScaler(feature_range=(0, 1))
#     X = minmax_scale.fit_transform(X)
    
    # 3-3 Label要轉成二進位的matrix，才能使用categorical_crossentropy
    from tensorflow.keras.utils import to_categorical
    y = to_categorical(Y)
    
    # 3-4 Train Test Split
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
    
    # 3-5 拿training data的10%作為Validation Split
    X_test, X_val, y_test, y_val = train_test_split(X_test, y_test, test_size=0.5, random_state=42)
    
    # 3-5 產生更多data
    for i in range(GenMoreData):
        X_train = np.append(X_train, X_train, axis=0)
        X_test = np.append(X_test, X_test, axis=0)
        X_val = np.append(X_val, X_val, axis=0)
        y_train = np.append(y_train, y_train, axis=0)
        y_test = np.append(y_test, y_test, axis=0)
        y_val = np.append(y_val, y_val, axis=0)
    
    return X_train, X_test, y_train, y_test, X_val, y_val

# 4 訓練MLP模型

import tensorflow as tf
from transformers import BertTokenizer, TFBertModel, TFBertForSequenceClassification, AdamW

from tensorflow import keras
from tensorflow.keras import Sequential, Model, Input
from tensorflow.keras.layers import Dense, Embedding, SpatialDropout1D, Dropout, Activation, Flatten, InputLayer
from sklearn import preprocessing
from sklearn.feature_selection import SelectFromModel
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, confusion_matrix
from sklearn.metrics import multilabel_confusion_matrix
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report

# # 如果GPU被占用
# config = tf.ConfigProto(
#         device_count = {'GPU': 0}
#     )
# session = tf.Session(config=config)

def MLP_model(INPUT_DIM, OUTPUT_DIM):
    model = Sequential()
    model.add(Dense(units=64, input_dim=INPUT_DIM, kernel_initializer='uniform', activation='relu'))
    model.add(Dense(units=32, activation='relu'))
    model.add(Dropout(0.2))
#     model.add(Dense(16, activation='relu'))
#     model.add(Dropout(0.2))
    model.add(Dense(units=OUTPUT_DIM, activation='softmax'))
    model.summary()
    
    return model

def show_train_history(train_history, train_key, val_key):
    plt.plot(train_history.history[train_key])
    plt.plot(train_history.history[val_key])
    plt.title('Model accuracy' if 'acc' in train_key else 'Model loss')
    plt.ylabel('accuracy' if 'acc' in train_key else 'loss')
    plt.xlabel('epoch')
    plt.legend(['train', 'validation'], loc='upper left')
    plt.plot()

X_train, X_test, y_train, y_test, X_val, y_val = dataPreprocessing(df_resample, IsSelected=False, HasCrisis=True, GenMoreData=6)

display(X_train.shape[1])

%%time
model = MLP_model(X_train.shape[1], y_train.shape[1])
# EPOCHS = 100
# LR = 0.1
# DECAY = LR / EPOCHS
# MOMENTUM = 0.8
# SGD = SGD(lr=LR, momentum=MOMENTUM, decay=DECAY, nesterov=False)
# ADAM = Adam(lr=LR, decay=DECAY)
model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['categorical_accuracy'])
history = model.fit(x=X_train, y=y_train, validation_data=(X_val, y_val), epochs=25, batch_size=16, verbose=2)

# model.save('model/model_TCRI.h5')

# tf.keras.models.save_model(
#     model,
#     "model/model_TCRI",
#     overwrite=True,
#     include_optimizer=True,
# ) 

# load_model = tf.keras.models.load_model('model/model_TCRI')
# load_model.summary()

X_train_RF, X_test_RF, y_train_RF, y_test_RF, X_val_RF, y_val_RF = dataPreprocessing(df_resample, IsSelected=False, HasCrisis=False, GenMoreData=6)

display(X_train_RF.shape[1])

%%time
model_RF = MLP_model(X_train_RF.shape[1], y_train_RF.shape[1])
# OPTIMIZER = Adam(lr=0.001)
model_RF.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['categorical_accuracy'])
history_RF = model_RF.fit(x=X_train_RF, y=y_train_RF, validation_data=(X_val_RF, y_val_RF), epochs=25, batch_size=16, verbose=2)

# model_RF.save('Model/TCRI/MLP_64units_32units_0513_withoutsent.h5')

# 5 評估模型與預測

## 5-1 載入模型

# # 載入先前已經訓練好的模型
# # model = load_model('Model/TCRI/MLP_64units_32units_0513_sentiment.h5')
# best_model = load_model('Model/TCRI/MLP_64units_32units_0513_sentiment.h5')

## 5-2 評估模型

### 5-2-1 有輿情

# # show_train_history(history, 'categorical_accuracy', 'val_categorical_accuracy')
# scores = model.evaluate(X_test, y_test)
# print(scores)

y_pred = model.predict_classes(X_test)
y_pred

y_test_classes = np.argmax(y_test, axis=1)
y_test_classes

# np.argmax(y_test, axis=1)

print(accuracy_score(y_test_classes, y_pred))
print(confusion_matrix(y_test_classes, y_pred))
print(classification_report(y_test_classes, y_pred))

# sns.heatmap(confusion_matrix(np.argmax(y_test, axis=1), y_pred), cmap='YlGnBu')

# pd.DataFrame(confusion_matrix(np.argmax(y_test, axis=1), y_pred)).to_csv('TCRI/ConfusionMatrix_Sentiment.csv')

### 5-2-2 沒有輿情

# 載入先前已經訓練好的模型
# model_RF = load_model('Model/TCRI/MLP_64units_32units_0513_withoutsent.h5')

# # show_train_history(history_RF, 'categorical_accuracy', 'val_categorical_accuracy')
# scores_RF = model_RF.evaluate(X_test_RF, y_test_RF)
# print(scores_RF)

y_pred_RF = model_RF.predict_classes(X_test_RF)
y_pred_RF

y_test_RF_classes = np.argmax(y_test_RF, axis=1)
y_test_RF_classes

print(accuracy_score(y_test_RF_classes, y_pred_RF))
print(confusion_matrix(y_test_RF_classes, y_pred_RF))
print(classification_report(y_test_RF_classes, y_pred_RF))

# sns.heatmap(confusion_matrix(np.argmax(y_test_RF, axis=1), y_pred_RF), cmap='YlGnBu')

# pd.DataFrame(confusion_matrix(np.argmax(y_test_RF, axis=1), y_pred_RF)).to_csv('TCRI/ConfusionMatrix_WithoutSent.csv')
